+++ packages/functions/src/dal/AppointmentDataAccessLayer.ts
@@ -26,6 +26,25 @@ export async function getAppointmentsByIds(
   return _.flatMap(snapshots, (s) => toDbAppointments(s));
 }
 
+export async function getAppointmentById(
+  appointmentId: string
+): Promise<DbAppointment | undefined> {
+  const foundAppointments = await getAppointmentsByIds([appointmentId]);
+  return foundAppointments[0];
+}
+
+export async function getAppointmentByIdOrThrow(
+  appointmentId: string
+): Promise<DbAppointment> {
+  const appointment = await getAppointmentById(appointmentId);
+
+  if (!appointment) {
+    throw new Error("Appointment not found");
+  }
+
+  return appointment;
+}
+
 export async function getAppointmentsCreatedByUserId(userId: string) {
   const appointments = (await admin
     .firestore()
